# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Billetterie Événementielle",
        "description": "Cette API permet de gérer les événements et les tickets pour une billetterie événementielle.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Afficher tous les événements",
                "operationId": "e529c81e6aaec82f49950a87ed881f57",
                "responses": {
                    "200": {
                        "description": "Une liste d'événements"
                    }
                }
            }
        },
        "/api/events/{id}/ticket-types": {
            "get": {
                "tags": [
                    "Ticket Types"
                ],
                "summary": "Get all ticket types for an event",
                "operationId": "f19a6be8ffe9abcb623ce49f9f91d3c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Events",
            "description": "Events"
        },
        {
            "name": "Ticket Types",
            "description": "Ticket Types"
        }
    ]
}